pyproject.toml
[project]
name = "tiled-acquisition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.4",
    "pymmcore-plus>=0.13.5",
    "useq-schema>=0.7.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
tiled_acquisition = "tiled_acquisition.main:main"